:truck: chore: modify , refactor,  update UI
-modify: file names and folder

- refactor: add form both payments and products

- update: create new UI for payment form
