+++ src/components/footer/Footer.module.scss
@@ -4,7 +4,7 @@
   padding: 1.5rem 0;
   color: rgb(40, 40, 40);
   border-top: 1.5px solid rgb(40, 40, 40);
-  margin: 0 3em;
+  margin: 2em 3em;
   .info, .links {
     margin: 2.9em 0;
   }
