+++ src/components/infoBox/InfoBox.scss
@@ -12,11 +12,7 @@
   color: #fff;
   transform: translateY(0);
   transition: all 0.3s;
-
-  // background-color: green;
-
   &:hover {
-    // border: 3px solid #fff;
     transform: translateY(-5px);
   }
 
@@ -27,13 +23,6 @@
   .info-text > * {
     color: #fff;
   }
-
-  // .card {
-  //   border: 1px solid #ccc;
-  //   border-bottom: 3px solid var(--light-blue);
-  //   padding: 5px;
-  //   background-color: #f5f6fa;
-  // }
 }
 
 @media screen and (max-width: 600px) {
