+++ src/data/sidebar.jsx
@@ -1,9 +1,7 @@
-import { RiAddCircleLine } from "react-icons/ri";
 import { BsPersonCircle } from "react-icons/bs";
 import { TbDatabase } from "react-icons/tb";
 import { MdOutlineMail, MdOutlinePayments } from "react-icons/md";
-import { CiImport } from "react-icons/ci"
-
+// import { CiImport } from "react-icons/ci"
 
 const menu = [
   // {
@@ -11,41 +9,21 @@ const menu = [
   //   icon: <FaTh />,
   //   path: "/dashboard",
   // },
-  // {
-  //   title: "Vendas",
-  //   icon: <MdOutlineAttachMoney />,
-  //   path: "/sales",
-  // },
   {
     title: "Estoque",
     icon: <TbDatabase />,
     path: "/storage",
   },
-  {
-    title: "Adicionar",
-    icon: <RiAddCircleLine />,
-    path: "/add-product",
-      //   childrens: [
-      // {
-      //   title: "Venda",
-      //   path: "/profile",
-      // },
-      // {
-      //   title: "Produto",
-      //   path: "/edit-profile",
-      // },
-    // ],
-  },
+  // {
+  //   title: "Caixa",
+  //   icon: <MdOutlineAttachMoney />,
+  //   path: "/sales",
+  // },
   {
     title: "Pagamentos",
     icon: <MdOutlinePayments />,
     path: "/payments",
   },
-  //   {
-  //   title: "Importar/Exportar",
-  //   icon: <MdOutlinePayments />,
-  //   path: "/payments",
-  // },
   // {
   //   title: "Importar",
   //   icon: <CiImport />,
@@ -62,5 +40,16 @@ const menu = [
     path: "/contact-us",
   },
 ];
+
+//   childrens: [
+// {
+//   title: "Venda",
+//   path: "/profile",
+// },
+// {
+//   title: "Produto",
+//   path: "/edit-profile",
+// },
+// ],
 
 export default menu;
